feat(GAT-6526): Homepage design tweaks (#1103)
