+++ src/app/[locale]/components/InfoHoverPanel/index.ts
