--- src/components/FlashyText/index.ts
@@ -1,3 +1,0 @@
-import FlashyText from "./FlashyText";
-
-export default FlashyText;
