+++ src/config/messages/en.json
@@ -805,7 +805,6 @@
             }
         },
         "home": {
-            "flashyText": "Explore the different functionalities of the Gateway",
             "touchDevice": "Explore the different functionalities of the Gateway",
             "seeNow": "See now",
             "meetTheTeam": "Meet the Team",
